+++ R/convertSimpleCountries.R
@@ -1,27 +1,31 @@
 #' @keywords datagen
-#' @export
-#' @title Convert World Borders Shapefile
+#' @title Convert (Simple) World Borders Shapefile
 #' @param nameOnly logical specifying whether to only return the name without creating the file
-#' @description Returns a SpatialPolygonsDataFrame for world divisions
+#' @description Returns a SpatialPolygonsDataFrame for a simple world divisions
 #' @details A world borders shapefile is downloaded and converted to a
 #' SpatialPolygonsDataFrame with additional columns of data. The resulting file will be created
-#' in the spatial data directory which is set with \code{setSpatialDataDir()}.
+#' in the package \code{SpatialDataDir} which is set with \code{setSpatialDataDir()}.
+#'
+#' This shapefile is a greatly simplified version of the TMWorldBorders shapefile and is especially suited
+#' for spatial searches. This is the default dataset used in \code{getCountry()} and \code{getCountryCode()}.
+#' Users may wish to use a higher resolution dataset when plotting.
+#' @note This is a non-exported function used only for updating the package dataset.
 #' @return Name of the dataset being created.
 #' @references \url{http://thematicmapping.org/downloads/}
 #' @seealso setSpatialDataDir
 #' @seealso getCountry, getCountryCode
-convertTMWorldBorders <- function(nameOnly=FALSE) {
+convertSimpleCountries <- function(nameOnly=FALSE) {
 
   # Use package internal data directory
   dataDir <- getSpatialDataDir()
 
   # Specify the name of the dataset and file being created
-  datasetName <- 'TMWorldBorders'
+  datasetName <- 'SimpleCountries'
 
   if (nameOnly) return(datasetName)
 
   # Build appropriate request URL for TM World Borders data
-  url <- 'http://thematicmapping.org/downloads/TM_WORLD_BORDERS-0.3.zip'
+  url <- 'http://thematicmapping.org/downloads/TM_WORLD_BORDERS_SIMPL-0.3.zip'
 
   filePath <- paste(dataDir,basename(url),sep='/')
   utils::download.file(url,filePath)
@@ -31,7 +35,7 @@ convertTMWorldBorders <- function(nameOnly=FALSE) {
   # Convert shapefile into SpatialPolygonsDataFrame
   # NOTE:  The 'world' directory has been created
   dsnPath <- paste(dataDir,'world',sep='/')
-  shpName <- 'TM_WORLD_BORDERS-0.3'
+  shpName <- 'TM_WORLD_BORDERS_SIMPL-0.3'
   SPDF <- convertLayer(dsn=dsnPath,layerName=shpName)
 
   # Rationalize naming:
@@ -49,6 +53,10 @@ convertTMWorldBorders <- function(nameOnly=FALSE) {
                    'area','population2005','UN_region','UN_subregion',
                    'longitude','latitude')
   
+  # NOTE:  http://conjugateprior.org/2013/01/unicode-in-r-packages-not/
+  # Transliterate unicode characters for this package-internal dataset
+  SPDF$countryName <- iconv(SPDF$countryName, from="UTF-8", to="ASCII//TRANSLIT")
+
   # Rationalize units:
   # * SI
   # NOTE:  Area seems to be in units of (10 km^2). Convert these to m^2
@@ -57,7 +65,7 @@ convertTMWorldBorders <- function(nameOnly=FALSE) {
   # Group polygons with the same identifier (countryCode)
   SPDF <- organizePolygons(SPDF, uniqueID='countryCode', sumColumns=c('area','population2005'))
   # NOTE:  This dataset already has grouped polygons
-  
+
   # Assign a name and save the data
   assign(datasetName,SPDF)
   save(list=c(datasetName),file=paste0(dataDir,'/',datasetName,'.RData'))